backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic and validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Google Cloud Platform
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-kms==2.18.0
google-auth==2.23.4

# AI and ML
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-google-genai==0.0.5

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
